refactor patient/doctor maps into separate class
